First Tack at kthelemn.hs
I am not doing this right. It seems like Haskell goes about this kind
of construction a bit differently than OCaml. Though in OCaml you
wouldn’t use the if then else syntax at all. Which it might not be a
bad idea to write an if-then-else function just to make these files a
little more concise…