Update to Pascal.py
Fixed an off by one error in the horizontal movement. Note, that if we
move too far to the right there is now an error because None has no
rikin or lefkin method! The fix is simple: use the symmetry of Pascal’s
triangle to make it so that if we move past the half way point of any
line, we just move back to the correct node. That way, we don’t have to
build a new bunch of crap into the data structure. Alternatively, you
could make it so that there’s a NullNode type, but this seems to be a
little simpler, though admittedly a little less extensible (not that
that’d ever really be relevant for this problem).