Funcin Around
Wish I knew more Java. But I’m learning a bit. I figured out that in
Scala, classes are a sort of function, at least they look analogous in
their declaration, I don’t know about their internal representation.
Now, because of strong typing this is clearly not right, since we would
want to define, say, the NullNode or the InitialNode using
PascalNode(0,0), but those are integers not PascalNodes! Hence, an
issue, that is also itself a question. Are classes the same as types in
Scala? One would kind of think so. Regardless, there ought to be a
mechanism like the variant types of OCaml to express this kind of
thing, so that we could define a type that could either be PascalNode
or Integer. They difficult part, is that we need to have a PascalNode
type to define that! Or at least, so it seems now. Time to watch some
lectures and stuff I guess.