collectrep.ml
Used exercise answer. Don’t understand it too well I’m afraid. I just
can’t quite figure out what the in clause is for. It must be because
we’re adding all of our collected lists on backwards, so then we feed
it into reverse… but how that ‘in’ works and why I do not know.