Need BufferedIO
Need to reread the buffo package. Basically, I’m not using it right. I
think you need to read things into the buffer (e.g. the whole file or
the file piecewise) and then operate on THAT as a string. You can’t
invoke range directly on a reader. That makes sense, as that would be
stupid. A reader isn’t a data set, it’s a method for getting a data set.