Coming along nicely
Issues: There’s something wrong with the le and ge methods on
CallNumber that’s making us exceed recursion depth when we do the list
comprehensions in quick sort. This is probably a super easy logic thing
but I’m tired and I can’t really fix on it right now.